修改Home界面布局

+++ src/components/Home.vue
@@ -2,22 +2,18 @@
 	
 	
 	
-	<el-container>
-			<!-- 头部 -->
+<!-- 	<el-container>
 			<el-header>
 			  <div>
 			    <img src="../assets/logo.png" alt />
 			    <span>苏宁易购后台管理系统</span>
 			  </div>
-			  <el-button type="danger" @click="logout">退出登录</el-button>
+			  <el-button type="infor" @click="logout">退出登录</el-button>
 			</el-header>
 		    
-			<!-- 主体 -->
 			<el-container>
-				<!-- 侧边栏 -->
 				<el-aside :width="isCollapse ? '64px' : '200px'">
 				
-				<div class="toggle-button" @click="isCollapse = !isCollapse"><i :class="isCollapse?'el-icon-d-arrow-right':'el-icon-d-arrow-left'"></i></div>
 				
 				 <el-menu 
 				 unique-opened 
@@ -28,14 +24,13 @@
 				 
 				  >
 		
-				   <!-- 一级菜单  -->
 				   <el-submenu :index="item.id+''" v-for="item in menuList" :key="item.id">
-				     <!-- 一级菜单的模板区域 -->
 				     <template slot="title">
-				       <i :class="iconObj[item.id]"></i>
-				       <span>{{ item.authName}}</span>
+						 <i :class="iconObj[item.id]" class="icos"></i>
+				       <span> 
+					   {{ item.authName}}
+					   </span>
 				     </template>
-				     <!-- 二级菜单 -->
 				     <el-menu-item 
 					 :index="'/' + subItem.path"
 					  v-for="subItem in item.children" 
@@ -43,9 +38,7 @@
 					   @click="saveNavState('/' + subItem.path)">
 					   
 					   
-				       <!-- 导航开启路由模式：
-				         将index值作为导航路由 -->
-				       <!-- 二级菜单的模板区域 -->
+				  
 				       <template slot="title">
 				         <i class="iconfont icon-shuzhuangtu"></i>
 				         <span>{{ subItem.authName}}</span>
@@ -57,7 +50,9 @@
 				      
 				
 				
-				
+				<div class="toggle-button" @click="isCollapse = !isCollapse">
+					<i :class="isCollapse?'el-icon-d-arrow-right':'el-icon-d-arrow-left'"></i>
+				</div>
 				
 				
 				
@@ -68,16 +63,121 @@
 				</el-aside>
 				
 				
-				<!-- 首页-->
 				<el-main>
 				  <router-view></router-view>
 				</el-main>
 				
 			</el-container>
+		</el-container> -->
+		
+		
+		<!-- 总体布局改造 -->
+		<el-container>
+		  <el-aside :width="isCollapse ? '64px' : '240px'">
+		  
+		  <el-menu
+		  unique-opened
+		  :collapse="isCollapse"
+		   :collapse-transition="false" 
+		   :default-active="activePath" 
+		  >
+			 <div :width="isCollapse ? '64px' : '240px'">
+				 <img src="../assets/logo.png" alt />
+			   <span v-if="!isCollapse">苏宁易购后台管理系统</span>
+			 </div>
+		  </el-menu>
+		  
+		   <el-menu 
+		   unique-opened 
+		   :collapse="isCollapse"
+		    :collapse-transition="false" 
+		    router 
+		    :default-active="activePath" 
+		   
+		    >
+		  	
+			
+				
+		     <el-submenu :index="item.id+''" v-for="item in menuList" :key="item.id">
+		       <template slot="title">
+		  		 <i :class="iconObj[item.id]" class="icos"></i>
+		         <span> 
+		  	   {{ item.authName}}
+		  	   </span>
+		       </template>
+		       <el-menu-item 
+		  	 :index="'/' + subItem.path"
+		  	  v-for="subItem in item.children" 
+		  	  :key="subItem.id"
+		  	   @click="saveNavState('/' + subItem.path)">
+		  	   
+		  	   
+		    
+		         <template slot="title">
+		           <i class="iconfont icon-shuzhuangtu"></i>
+		           <span>{{ subItem.authName}}</span>
+		         </template>
+		       </el-menu-item>
+		     </el-submenu>
+		   </el-menu>
+		  		
+		        
+		  
+		  
+		  <div class="toggle-button" @click="isCollapse = !isCollapse">
+		  	<i :class="isCollapse?'el-icon-d-arrow-right':'el-icon-d-arrow-left'"></i>
+		  </div>
+		  
+		  
+		  
+		  
+		  
+		  
+		  
+		  </el-aside>
+		  <!-- <el-container>
+			  <el-header>
+			  <el-button type="infor" @click="logout">退出登录</el-button>
+			  
+			  </el-header>
+		    <el-main>
+				
+				
+					
+				<router-view></router-view>
+			</el-main>
+		  </el-container> -->
+		  <el-container>
+		      <el-header>
+			  
+			  <el-row>
+				  <el-col class="header-right">
+				    <el-dropdown @command="handerCommand">
+				      <span>
+				        欢迎您 admin
+				        <i class="el-icon-arrow-down el-icon--right"></i>
+				      </span>
+				      <el-dropdown-menu slot="dropdown">
+				        <el-dropdown-item command="home">个人中心</el-dropdown-item>
+				        <el-dropdown-item command="logout">退出登录</el-dropdown-item>
+				      </el-dropdown-menu>
+				    </el-dropdown>
+				    <el-avatar src="https://cube.elemecdn.com/0/88/03b0d39583f48206768a7534e55bcpng.png">
+						
+					</el-avatar>
+				  </el-col>
+			  </el-row>
+			  <!-- <el-button type="infor" @click="logout">退出登录</el-button> -->
+			  </el-header>
+		      <el-main>
+				 <router-view></router-view> 
+			  </el-main>
+		    </el-container>
 		</el-container>
 		
 		
-	</el-container>
+		
+		
 </template>
 
 <script>
@@ -96,7 +196,7 @@
 			  // 默认不折叠
 			  isCollapse: false,
 			  // 被激活导航地址
-			  activePath: ''
+			  activePath: '',
 			}
 		},
 		created () {
@@ -104,11 +204,27 @@
 		  this.activePath = window.sessionStorage.getItem('activePath')
 		},
 		methods: {
-		  logout () {
-		    // 清空token
-		    window.sessionStorage.clear()
-		    this.$router.push('/login')
-		  },
+			handerCommand(cmd) {
+			  console.log(cmd);
+			  if (cmd === "home") {
+			    // 去到home页面
+			    this.$router.push("/home");
+			  } else {
+			    this.$message.warning("您已成功退出!!");
+			    // 清空token
+			    window.sessionStorage.clear()
+			    this.$router.push('/login')
+			    // 刷新浏览器【重新初始化路由】
+			    location.reload();
+			  }
+			},
+			
+			
+		  // logout () {
+		  //   // 清空token
+		  //   window.sessionStorage.clear()
+		  //   this.$router.push('/login')
+		  // },
 		  // 获取请求菜单
 		  async getMenuList () {
 		    const { data: res } = await this.$http.get('menus')
@@ -130,52 +246,77 @@
 	  height: 100%;
 	}
 	.el-header {
-	  background-color: #0018bb;
+	  background-color: #f8f9fc;
+	  box-shadow:inset 0px 0px 20px 20px #e9eaed;
+	  
 	  display: flex;
-	  justify-content: space-between;
-	  padding-left: 0;
+	  justify-content:flex-end;
 	  align-items: center;
 	  color: #fff;
 	  font-size: 20px;
-	  > div {
-	    display: flex;
-	    align-items: center;
-	    img {
-	      height: 40px;
-	    }
-	    span {
-	      margin-left: 15px;
-	    }
+	  .el-dropdown{
+		font-size: 20px;  
 	  }
 	}
 	.el-aside {
-	  background-color: #2b3140;
+	  background:linear-gradient(to bottom,#4e73df,#3d63d2);
 	
-	  // .el-menu {
-	  //   border: none;
-	  // }
+	  .el-menu {
+	    border: none;
+		
+		> div {
+		  display: flex;
+		  align-items: center;
+		  overflow: hidden;
+		  flex-wrap: nowrap;
+		  font-weight: 900;
+		  background:linear-gradient(to bottom,#4e73df,#3d63d2);
+		  color: #fff;
+		  img {
+		    height: 56px;
+		  }
+		  span {
+		    margin-left: 10px;
+		  }
+		}
+		
+		
+		
+	  }
 	  .el-menu-item,
 	  /deep/ .el-submenu__title {
-	    color: #fff;
-		background-color: #1c212b;
+	    color: rgba(255,255,255,0.6);
+		background:linear-gradient(to bottom,#4e73df,#3d63d2);
+		  
 	  }
 	  /deep/ .el-menu-item:focus,
 	  /deep/ .el-menu-item:hover,
 	  /deep/.el-submenu__title:hover {
 	    outline: 0;
-		color: #0055ff;
+		color: #fff;
 		background-color: rgb(49,83,120);
+		.icos{
+			color: #fff;
+		}
 	  }
 	  /deep/ .el-menu-item.is-active {
-	    color: #ffaa00;
-		background-color:#e5e9ef;
+	    color: #fff;
+		background-color:#ffffff;
+		border-radius: 10px;
 	  }
 	  
+	  .icos{
+		  color: rgba(255,255,255,0.6);
+		  &:hover{
+			  color: #fff;
+		  }
+	  }
 	  
 	  
 	}
 	.el-main {
-	  background-color: #e5e9ef;
+	  background-color: #f8f9fc;
+	  box-shadow:inset 0px 0px 20px 20px #e9eaed;
 	}
 	
 	.iconfont{
@@ -184,22 +325,30 @@
 	}
 	
 	.toggle-button {
-	  background-color: #29a1b1;
-	  font-size: 10px;
-	  line-height: 24px;
-	  color: #fff;
+	  width: 50px;
+	  height: 50px;
+	  border-radius: 50%;
+	  background: rgb(102, 133, 221);
+	  color: rgb(181, 196, 240);
 	  text-align: center;
-	  letter-spacing: 0.2em;
+	  line-height: 50px;
+	  opacity: 0.8;
+	  margin: 50px auto;
+	  
 	 &:hover{
 		  cursor: pointer;
-		  background: #ff5d0d;
-		  color: white;
+		  opacity: 1;
+	 }
+	 i{
+		 font-weight: 900;
 	 }
 	}
 	.el-icon-d-arrow-right::before{
 		font-size: 16px;
+		font-weight: 900;
 	}
 	.el-icon-d-arrow-left::before{
 		font-size: 16px;
+		font-weight: 900;
 	}
 </style>
